chore(toolsets): move tests to another package
